+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - [upgrade redisgo to v1.8.8](https://github.com/beego/beego/pull/4872)
 - [fix prometheus CVE-2022-21698](https://github.com/beego/beego/pull/4878)
 - [upgrade to Go 1.18](https://github.com/beego/beego/pull/4896)
+- [Add httplib OpenTelemetry Filter](https://github.com/beego/beego/pull/4888)
 
 # v2.0.2
 See v2.0.2-beta.1